refactor: move outsider_events into relationship with other event files
