+++ scripts/events_module/relationship/outsider_events.py
